H3DAPI warning fixes (vs2010)
Summary: Changes to various things in H3DAPI giving warnings in vs2010

Test Plan: Need to figure one out

Reviewers: markus

Reviewed By: markus

Subscribers: markus, michaels

Tags: #h3dapi_2.x

Differential Revision: https://www.h3dapi.org:4000/D29

+++ H3DAPI/src/PythonScript.cpp
@@ -395,14 +395,14 @@ void PythonScript::loadScript( const string &script_filename, const string &scri
     
       // get length of file:
       is.seekg (0, ios::end);
-      length = is.tellg();
+      length = static_cast<int>(is.tellg());
       is.seekg (0, ios::beg);
     
       // allocate memory:
       buffer = new char [length + 1];
       // read data as a block:
       is.read (buffer,length);
-      length = is.gcount();
+      length = static_cast<int>(is.gcount());
       is.close();
       buffer[length ] = '\0';
